refactor: add useCalendarModal hook
Updated Calendar component to use useCalendarModal hook to handle
calendar visibility and move the logic from the component
